fix landing Page
