+++ src/components/taskPage/maincontent.tsx
@@ -4,16 +4,17 @@ import SectionTitle from "../font/sectionTitle";
 import LogoHeader from "../logoHeader";
 import SearchBar from "../searchBar";
 import CTA from "../common/cta";
-import {JobListPage} from "../home/JobCard";
+import { JobListPage } from "../home/JobCard";
 import {
     Box,
     Typography,
     Chip,
     Divider,
 } from "@mui/material";
-import {Job} from "enigma/types/models";
-import {useSearchParams} from "next/navigation";
-import {useRouter} from "next/navigation";
+import { Job } from "enigma/types/models";
+import { useSearchParams } from "next/navigation";
+import { useRouter } from "next/navigation";
+import LoadingData from "./loadingData";
 
 export const MainContent = () => {
     const router = useRouter();
@@ -26,96 +27,18 @@ export const MainContent = () => {
         "Hotel Receptionist",
     ];
 
-    const [jobs, setJobs] = React.useState<Job[]>([]);
-    const [loading, setLoading] = React.useState(true);
-
-    React.useEffect(() => {
-        const fetchJobs = async () => {
-            setLoading(true)
-            try {
-                const queryParams = new URLSearchParams();
-                queryParams.set('status', 'active,prioritized');
-                const query = searchParams.get('query');
-                if (query) {
-                    queryParams.set('query', query);
-                }
-
-                const locations = searchParams.get('locations');
-                if (locations) {
-                    queryParams.set('locations', locations);
-                }
-
-                const jobFunctions = searchParams.get('jobFunctions');
-                if (jobFunctions) {
-                    queryParams.set('jobFunctions', jobFunctions);
-                }
-
-                const jobSubfunctions = searchParams.get('jobSubfunctions');
-                if (jobSubfunctions) {
-                    queryParams.set('jobSubfunctions', jobSubfunctions);
-                }
-
-                const industries = searchParams.get('industries');
-                if (industries) {
-                    queryParams.set('industries', industries);
-                }
-
-                const employmentType = searchParams.get('employment_type');
-                if (employmentType) {
-                    queryParams.set('employment_type', employmentType);
-                }
-
-                //TODO: other filter criteria will continue from here
-
-                const page = searchParams.get('page') || '1';
-                queryParams.set('page', page);
-
-                console.log('query params', queryParams.toString());
-
-                const response = await fetch(`/api/jobs?${queryParams.toString()}`);
-                if (!response.ok) {
-                    throw new Error('failed to fetch jobs');
-                }
-
-                const data = await response.json();
-                if (data.jobs) {
-                    // Make sure each job has the required properties
-                    const transformedJobs = data.jobs.map((job: Job) => ({
-                        ...job,
-                        // Ensure the job has industry object with industry_name
-                        industry: job.industry || {industry_name: ""},
-                        // Convert any string dates to Date objects if needed
-                        close_date: job.close_date ? new Date(job.close_date) : new Date()
-                    }));
-                    setJobs(transformedJobs);
-                } else {
-                    console.error('the response have unexpected data', data);
-                    setJobs([]);
-                }
-            } catch (error) {
-                console.error("job fetch failed: ", error);
-                setJobs([]);
-            } finally {
-                setLoading(false);
-            }
-        };
-        fetchJobs();
-    }, [searchParams]);
-
-    console.log(`FETCHED JOBS: `, jobs);
-
     return (
         <Box component="main" sx={{
             flexGrow: 1,
-            p: {xs: 0.5, sm: 3},
+            p: { xs: 0.5, sm: 3 },
             width: '100%',
             ml: 0.5,
             '@media (max-width: 991px)': {
                 maxWidth: '100%',
             },
         }}>
-            <Box sx={{display: {lg: 'none', sm: 'block'}}}>
-                <Divider sx={{mt: 1, mb: 3, width: '100%'}}/>
+            <Box sx={{ display: { lg: 'none', sm: 'block' } }}>
+                <Divider sx={{ mt: 1, mb: 3, width: '100%' }} />
             </Box>
 
             <Typography variant="h4" component="h1" gutterBottom color="#101828">
@@ -123,7 +46,7 @@ export const MainContent = () => {
             </Typography>
 
             {/* Search and Filter */}
-            <SearchBar placeholder="Job title, description, or industry name"/>
+            <SearchBar placeholder="Job title, description, or industry name" />
 
 
             {/* ADDED TO CHECK THE ACTIVATED FILTER */}
@@ -134,85 +57,85 @@ export const MainContent = () => {
                 searchParams.get('industries') ||
                 searchParams.get('employment_type')
             ) && (
-                <Box sx={{mb: 2, p: 2, bgcolor: '#f5f5f5', borderRadius: 1}}>
-                    <Typography variant="body2" color="text.secondary" sx={{mb: 1}}>
-                        Active Filters:
-                    </Typography>
-                    <Box sx={{display: 'flex', gap: 1, flexWrap: 'wrap'}}>
-                        {searchParams.get('query') && (
-                            <Chip
-                                label={`Search: ${searchParams.get('query')}`}
-                                size="small"
-                                onDelete={() => {
-                                    const newParams = new URLSearchParams(searchParams.toString());
-                                    newParams.delete('query');
-                                    router.push(`/jobs?${newParams.toString()}`);
-                                }}
-                            />
-                        )}
-                        {searchParams.get('locations') && (
-                            <Chip
-                                label={`Locations: ${searchParams.get('locations')?.replace(/,/g, ', ')}`}
-                                size="small"
-                                onDelete={() => {
-                                    const newParams = new URLSearchParams(searchParams.toString());
-                                    newParams.delete('locations');
-                                    router.push(`/jobs?${newParams.toString()}`);
-                                }}
-                            />
-                        )}
-                        {/* Add other filter chips... */}
-                        {searchParams.get('jobFunctions') && (
-                            <Chip
-                                label={`Job Functions: ${searchParams.get('jobFunctions')?.replace(/,/g, ', ')}`}
-                                size="small"
-                                onDelete={() => {
-                                    const newParams = new URLSearchParams(searchParams.toString());
-                                    newParams.delete('jobFunctions');
-                                    router.push(`/jobs?${newParams.toString()}`);
-                                }}
-                            />
-                        )}
+                    <Box sx={{ mb: 2, p: 2, bgcolor: '#f5f5f5', borderRadius: 1 }}>
+                        <Typography variant="body2" color="text.secondary" sx={{ mb: 1 }}>
+                            Active Filters:
+                        </Typography>
+                        <Box sx={{ display: 'flex', gap: 1, flexWrap: 'wrap' }}>
+                            {searchParams.get('query') && (
+                                <Chip
+                                    label={`Search: ${searchParams.get('query')}`}
+                                    size="small"
+                                    onDelete={() => {
+                                        const newParams = new URLSearchParams(searchParams.toString());
+                                        newParams.delete('query');
+                                        router.push(`/jobs?${newParams.toString()}`);
+                                    }}
+                                />
+                            )}
+                            {searchParams.get('locations') && (
+                                <Chip
+                                    label={`Locations: ${searchParams.get('locations')?.replace(/,/g, ', ')}`}
+                                    size="small"
+                                    onDelete={() => {
+                                        const newParams = new URLSearchParams(searchParams.toString());
+                                        newParams.delete('locations');
+                                        router.push(`/jobs?${newParams.toString()}`);
+                                    }}
+                                />
+                            )}
+                            {/* Add other filter chips... */}
+                            {searchParams.get('jobFunctions') && (
+                                <Chip
+                                    label={`Job Functions: ${searchParams.get('jobFunctions')?.replace(/,/g, ', ')}`}
+                                    size="small"
+                                    onDelete={() => {
+                                        const newParams = new URLSearchParams(searchParams.toString());
+                                        newParams.delete('jobFunctions');
+                                        router.push(`/jobs?${newParams.toString()}`);
+                                    }}
+                                />
+                            )}
 
-                        {searchParams.get('jobSubfunctions') && (
-                            <Chip
-                                label={`Job Subfunctions: ${searchParams.get('jobSubfunctions')?.replace(/,/g, ', ')}`}
-                                size="small"
-                                onDelete={() => {
-                                    const newParams = new URLSearchParams(searchParams.toString());
-                                    newParams.delete('jobSubfunctions');
-                                    router.push(`/jobs?${newParams.toString()}`);
-                                }}
-                            />
-                        )}
+                            {searchParams.get('jobSubfunctions') && (
+                                <Chip
+                                    label={`Job Subfunctions: ${searchParams.get('jobSubfunctions')?.replace(/,/g, ', ')}`}
+                                    size="small"
+                                    onDelete={() => {
+                                        const newParams = new URLSearchParams(searchParams.toString());
+                                        newParams.delete('jobSubfunctions');
+                                        router.push(`/jobs?${newParams.toString()}`);
+                                    }}
+                                />
+                            )}
 
-                        {searchParams.get('industries') && (
-                            <Chip
-                                label={`Industries: ${searchParams.get('industries')?.replace(/,/g, ', ')}`}
-                                size="small"
-                                onDelete={() => {
-                                    const newParams = new URLSearchParams(searchParams.toString());
-                                    newParams.delete('industries');
-                                    router.push(`/jobs?${newParams.toString()}`);
-                                }}
-                            />
-                        )}
+                            {searchParams.get('industries') && (
+                                <Chip
+                                    label={`Industries: ${searchParams.get('industries')?.replace(/,/g, ', ')}`}
+                                    size="small"
+                                    onDelete={() => {
+                                        const newParams = new URLSearchParams(searchParams.toString());
+                                        newParams.delete('industries');
+                                        router.push(`/jobs?${newParams.toString()}`);
+                                    }}
+                                />
+                            )}
 
-                        {searchParams.get('employment_type') && (
-                            <Chip
-                                label={`Employment Type: ${searchParams.get('employment_type')?.replace(/,/g, ', ')}`}
-                                size="small"
-                                onDelete={() => {
-                                    const newParams = new URLSearchParams(searchParams.toString());
-                                    newParams.delete('employment_type');
-                                    router.push(`/jobs?${newParams.toString()}`);
-                                }}
-                            />
-                        )}
+                            {searchParams.get('employment_type') && (
+                                <Chip
+                                    label={`Employment Type: ${searchParams.get('employment_type')?.replace(/,/g, ', ')}`}
+                                    size="small"
+                                    onDelete={() => {
+                                        const newParams = new URLSearchParams(searchParams.toString());
+                                        newParams.delete('employment_type');
+                                        router.push(`/jobs?${newParams.toString()}`);
+                                    }}
+                                />
+                            )}
 
+                        </Box>
                     </Box>
-                </Box>
-            )}
+                )}
 
             {/* Popular Jobs */}
             <Box sx={{
@@ -225,10 +148,10 @@ export const MainContent = () => {
                 },
 
             }}>
-                <Typography variant="h6" color="text.secondary" sx={{mb: 1}}>
+                <Typography variant="h6" color="text.secondary" sx={{ mb: 1 }}>
                     Popular Jobs:
                 </Typography>
-                <Box sx={{display: "flex", gap: 1.5, flexWrap: "wrap"}}>
+                <Box sx={{ display: "flex", gap: 1.5, flexWrap: "wrap" }}>
                     {popularJobs.map((job, index) => (
                         <Chip
                             key={index}
@@ -241,13 +164,13 @@ export const MainContent = () => {
                                 '&:hover': {
                                     backgroundColor: '#2494B620' // 20% opacity
                                 }
-                            }}/>
+                            }} />
                     ))}
                 </Box>
             </Box>
 
             <Box sx={{
-                display: "flex", flexDirection: {xs: "column", lg: "row"}, gap: 4, width: '100%'
+                display: "flex", flexDirection: { xs: "column", lg: "row" }, gap: 4, width: '100%'
             }}>
                 {/* Left Column */}
                 <Box sx={{
@@ -258,23 +181,13 @@ export const MainContent = () => {
                     },
                 }}>
                     {/* Step Section */}
-                    <SectionTitle title="Step Into Your Future" showOptions/>
-                    <CTA/>
+                    <SectionTitle title="Step Into Your Future" showOptions />
+                    <CTA />
 
                     {/* Trending Jobs */}
-                    <SectionTitle title="Trending Jobs This Week" showOptions1 showOptions/>
+                    <SectionTitle title="Trending Jobs This Week" showOptions1 showOptions />
 
-                    {loading ? (
-                        <Box sx={{textAlign: 'center', py: 4}}>
-                            <Typography variant="body1">Loading jobs...</Typography>
-                        </Box>
-                    ) : jobs.length > 0 ? (
-                        <JobListPage jobs={jobs}/>
-                    ) : (
-                        <Box sx={{textAlign: 'center', py: 4}}>
-                            <Typography variant="body1">No jobs found matching your criteria.</Typography>
-                        </Box>
-                    )}
+                    <LoadingData />
 
                 </Box>
             </Box>
